feat: Adds support for ulong and uint

+++ Benchmarks/src/BoxingBenchmarks.cs
@@ -7,13 +7,14 @@ namespace Benchmarks;
 [SuppressMessage("ReSharper", "UnusedMember.Global")]
 [SuppressMessage("ReSharper", "UnusedType.Global")]
 public class BoxingBenchmarks {
-	public static int Iterations;
-	public static int LoopIterations;
+	public static ulong Iterations;
+	public static ulong LoopIterations;
+
 
 	[Benchmark("BoxedInteger", "Tests operation on boxed int")]
 	public static Int32 BInt() {
 		Int32 boxed = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			boxed++;
 			boxed *= 3;
 			boxed /= 2;
@@ -27,7 +28,7 @@ public class BoxingBenchmarks {
 	[Benchmark("BoxedInteger", "Tests operation on boxed uint")]
 	public static UInt32 BUint() {
 		UInt32 boxed = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			boxed++;
 			boxed *= 3;
 			boxed /= 2;
@@ -42,7 +43,7 @@ public class BoxingBenchmarks {
 	public static nint BNint() {
 		nint primitive = 0;
 		object boxed = primitive;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			boxed = (nint)boxed + 1;
 			boxed = (nint)boxed * 3;
 			boxed = (nint)boxed / 2;
@@ -57,7 +58,7 @@ public class BoxingBenchmarks {
 	public static nuint BNuint() {
 		nuint primitive = 0;
 		object boxed = primitive;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			boxed = (nuint)boxed + 1;
 			boxed = (nuint)boxed * 3;
 			boxed = (nuint)boxed / 2;
@@ -71,7 +72,7 @@ public class BoxingBenchmarks {
 	[Benchmark("BoxedInteger", "Tests operation on boxed long")]
 	public static Int64 BLong() {
 		Int64 boxed = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			boxed++;
 			boxed *= 3;
 			boxed /= 2;
@@ -85,7 +86,7 @@ public class BoxingBenchmarks {
 	[Benchmark("BoxedInteger", "Tests operation on boxed ulong")]
 	public static UInt64 BUlong() {
 		UInt64 boxed = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			boxed++;
 			boxed *= 3;
 			boxed /= 2;
@@ -99,7 +100,7 @@ public class BoxingBenchmarks {
 	[Benchmark("BoxedInteger", "Tests operation on boxed short")]
 	public static Int16 BShort() {
 		Int16 boxed = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			boxed = (short)(boxed + 1);
 			boxed = (short)(boxed * 3);
 			boxed = (short)(boxed / 2);
@@ -113,7 +114,7 @@ public class BoxingBenchmarks {
 	[Benchmark("BoxedInteger", "Tests operation on boxed ushort")]
 	public static UInt16 BUshort() {
 		UInt16 boxed = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			boxed = (ushort)(boxed + 1);
 			boxed = (ushort)(boxed * 3);
 			boxed = (ushort)(boxed / 2);
@@ -127,7 +128,7 @@ public class BoxingBenchmarks {
 	[Benchmark("BoxedInteger", "Tests operation on boxed byte")]
 	public static Byte BByte() {
 		Byte boxed = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			boxed = (byte)(boxed + 1);
 			boxed = (byte)(boxed * 3);
 			boxed = (byte)(boxed / 2);
@@ -141,7 +142,7 @@ public class BoxingBenchmarks {
 	[Benchmark("BoxedInteger", "Tests operation on boxed sbyte")]
 	public static SByte BSbyte() {
 		SByte boxed = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			boxed = (sbyte)(boxed + 1);
 			boxed = (sbyte)(boxed * 3);
 			boxed = (sbyte)(boxed / 2);
@@ -155,7 +156,7 @@ public class BoxingBenchmarks {
 	[Benchmark("BoxedDecimal", "Tests operation on boxed float")]
 	public static Single BFloat() {
 		Single boxed = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			boxed++;
 			boxed *= 3;
 			boxed /= 2;
@@ -169,7 +170,7 @@ public class BoxingBenchmarks {
 	[Benchmark("BoxedDecimal", "Tests operation on boxed double")]
 	public static Double BDouble() {
 		Double boxed = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			boxed++;
 			boxed *= 3;
 			boxed /= 2;
@@ -183,7 +184,7 @@ public class BoxingBenchmarks {
 	[Benchmark("BoxedDecimal", "Tests operation on boxed decimal")]
 	public static Decimal BDecimal() {
 		Decimal boxed = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			boxed++;
 			boxed *= 3;
 			boxed /= 2;
@@ -197,7 +198,7 @@ public class BoxingBenchmarks {
 	[Benchmark("BoxedBool", "Tests setting bool values")]
 	public static Boolean BBool() {
 		Boolean boxed = false;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			if (boxed) {
 				boxed = false;
 			}
@@ -212,7 +213,7 @@ public class BoxingBenchmarks {
 	[Benchmark("BoxedBool", "Tests using byte as boolean values")]
 	public static Byte BByteAsBool() {
 		Byte boxed = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			if (boxed != 0) {
 				boxed = 0;
 			}
