+++ Benchmarks/src/Collections/Set/HashSetBenchmarks.cs
@@ -7,8 +7,9 @@ namespace Benchmarks.Collections.Set;
 [SuppressMessage("ReSharper", "UnusedMember.Global")]
 [SuppressMessage("ReSharper", "UnusedType.Global")]
 public class HashSetBenchmarks {
-	public static int Iterations;
-	public static int LoopIterations;
+	public static ulong Iterations;
+	public static ulong LoopIterations;
+
 	public static readonly HashSet<int> Data = new(1000);
 
 
@@ -21,7 +22,7 @@ public class HashSetBenchmarks {
 	[Benchmark("SetCreation", "Tests allocation and initialization of a HashSet")]
 	public static int HashSetCreation() {
 		int result = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			HashSet<int> array = new HashSet<int>();
 			for (int index = 0; index < Data.Count; index++) {
 				array.Add(index * 2);
@@ -36,7 +37,7 @@ public class HashSetBenchmarks {
 	[Benchmark("SetGet", "Tests getting values from a HashSet")]
 	public static int HashSetGet() {
 		int sum = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			for (int j = 0; j < Data.Count; j++) {
 				Data.TryGetValue(j, out int value);
 				sum += value;
@@ -49,7 +50,7 @@ public class HashSetBenchmarks {
 	[Benchmark("SetInsertion", "Tests insertion into a HashSet")]
 	public static int HashSetInsertion() {
 		HashSet<int> temp = new();
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			temp = new HashSet<int>();
 			for (int j = 0; j < 1000; j++) {
 				temp.Add(j);
@@ -62,7 +63,7 @@ public class HashSetBenchmarks {
 	[Benchmark("SetRemoval", "Tests removal from a HashSet")]
 	public static int HashSetRemoval() {
 		HashSet<int> temp = new();
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			for (int j = 0; j < 1000; j++) {
 				temp.Add(j);
 			}
@@ -78,7 +79,7 @@ public class HashSetBenchmarks {
 	[Benchmark("SetCopy", "Tests copying a HashSet using a foreach loop")]
 	public static int HashSetCopyManualForeach() {
 		int result = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			HashSet<int> target = new HashSet<int>();
 			foreach (int element in Data) {
 				target.Add(element);
@@ -94,7 +95,7 @@ public class HashSetBenchmarks {
 	[Benchmark("SetCopy", "Tests copying a HashSet using a for loop")]
 	public static int HashSetCopyManualFor() {
 		int result = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			HashSet<int> target = new HashSet<int>();
 			for (int j = 0; j < Data.Count; j++) {
 				Data.TryGetValue(j, out int value);
