+++ Benchmarks/src/Collections/Table/DictionaryBenchmarks.cs
@@ -9,8 +9,9 @@ namespace Benchmarks.Collections.Table;
 [SuppressMessage("ReSharper", "UnusedMember.Global")]
 [SuppressMessage("ReSharper", "UnusedType.Global")]
 public class DictionaryBenchmarks {
-	public static int Iterations;
-	public static int LoopIterations;
+	public static ulong Iterations;
+	public static ulong LoopIterations;
+
 	public static readonly Dictionary<int, int> Data = new(1000);
 
 	static DictionaryBenchmarks() {
@@ -23,7 +24,7 @@ public class DictionaryBenchmarks {
 	[Benchmark("TableInsertion", "Tests insertion into a Dictionary")]
 	public static int DictionaryInsertion() {
 		Dictionary<int, int> temp = new Dictionary<int, int>();
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			temp = new Dictionary<int, int>();
 			for (int j = 0; j < 1000; j++) {
 				temp.Add(CollectionsHelpers.SequentialIndices[j], CollectionsHelpers.RandomValues[j]);
@@ -36,7 +37,7 @@ public class DictionaryBenchmarks {
 	[Benchmark("TableGet", "Tests getting values sequentially from a Dictionary")]
 	public static int DictionaryGet() {
 		int sum = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			for (int j = 0; j < Data.Count; j++) {
 				sum += Data[CollectionsHelpers.SequentialIndices[j]];
 			}
@@ -48,7 +49,7 @@ public class DictionaryBenchmarks {
 	[Benchmark("TableGet", "Tests getting values randomly from a Dictionary")]
 	public static int DictionaryGetRandom() {
 		int sum = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			for (int j = 0; j < Data.Count; j++) {
 				sum += Data[CollectionsHelpers.RandomIndices[j]];
 			}
@@ -60,7 +61,7 @@ public class DictionaryBenchmarks {
 	[Benchmark("TableRemoval", "Tests removal from a Dictionary")]
 	public static int DictionaryRemoval() {
 		Dictionary<int, int> temp = new Dictionary<int, int>();
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			for (int j = 0; j < 1000; j++) {
 				temp.Add(CollectionsHelpers.SequentialIndices[j], CollectionsHelpers.RandomValues[j]);
 			}
@@ -76,7 +77,7 @@ public class DictionaryBenchmarks {
 	[Benchmark("TableCreation", "Tests allocation and initialization of a Dictionary")]
 	public static int DictionaryCreation() {
 		int result = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			Dictionary<int, int> dictionary = new Dictionary<int, int>();
 			for (int index = 0; index < Data.Count; index++) {
 				dictionary.Add(index, index * 2);
@@ -91,7 +92,7 @@ public class DictionaryBenchmarks {
 	[Benchmark("TableCopy", "Tests copying a Dictionary using a foreach loop looping through k/v pairs")]
 	public static int DictionaryCopyManualForeach() {
 		int result = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			Dictionary<int, int> target = new Dictionary<int, int>();
 
 			foreach (KeyValuePair<int, int> pair in Data) {
@@ -108,7 +109,7 @@ public class DictionaryBenchmarks {
 	[Benchmark("TableCopy", "Tests copying a Dictionary using a foreach loop and looping through keys")]
 	public static int DictionaryCopyManualForeachIndex() {
 		int result = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			Dictionary<int, int> target = new Dictionary<int, int>();
 
 			foreach (int key in Data.Keys) {
@@ -125,7 +126,7 @@ public class DictionaryBenchmarks {
 	[Benchmark("TableCopy", "Tests copying a Dictionary using a foreach loop using deconstruction")]
 	public static int DictionaryCopyManualForeachDeconstruct() {
 		int result = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			Dictionary<int, int> target = new Dictionary<int, int>();
 
 			foreach ((int key, int value) in Data) {
